hf_space/requirements.txt
torch==2.1.0
torchvision==0.16.0
gradio==4.44.1
Pillow==10.1.0
timm==0.9.12
numpy==1.24.4
